+++ pgpy/packet/packets.py
@@ -394,7 +394,7 @@ class SignatureV4(Signature):
         self._signature = val
 
     @property
-    def signer(self):
+    def signer(self) -> KeyID:
         return self.subpackets['Issuer'][-1].issuer
 
     def __init__(self):
@@ -702,23 +702,19 @@ class OnePassSignatureV3(OnePassSignature):
             self._halg = val
 
     @sdproperty
-    def signer(self):
+    def signer(self) -> KeyID:
         return self._signer
 
-    @signer.register(str)
-    def signer_str(self, val):
-        self._signer = val
-
-    @signer.register(bytearray)
-    def signer_bin(self, val):
-        self._signer = binascii.hexlify(val).upper().decode('latin-1')
+    @signer.register
+    def signer_bin(self, val:Union[bytearray,bytes,str,KeyID,Fingerprint]):
+        self._signer = KeyID(val)
 
     def __init__(self):
         super(OnePassSignatureV3, self).__init__()
         self._sigtype = None
         self._halg = None
         self._pubalg = None
-        self._signer = b'\x00' * 8
+        self._signer = KeyID(b'\x00'*8)
         self.nested = False
 
     def __bytearray__(self):
@@ -727,7 +723,7 @@ class OnePassSignatureV3(OnePassSignature):
         _bytes += bytearray([self.sigtype])
         _bytes += bytearray([self.halg])
         _bytes += bytearray([self.pubalg])
-        _bytes += binascii.unhexlify(self.signer.encode("latin-1"))
+        _bytes += bytes(self.signer)
         _bytes += bytearray([int(self.nested)])
         return _bytes