+++ pgpy/pgp.py
@@ -287,7 +287,7 @@ class PGPSignature(Armorable, ParentRef, PGPObject):
         return ret
 
     @property
-    def signer(self):
+    def signer(self) -> Optional[KeyID]:
         """
         The 16-character Key ID of the key that generated this signature.
         """
@@ -909,8 +909,8 @@ class PGPMessage(Armorable, PGPObject):
         return len(self._signatures) > 0
 
     @property
-    def issuers(self):
-        """A ``set`` containing all key ids (if any) which have signed or encrypted this message."""
+    def issuers(self) -> Set[Union[KeyID,Fingerprint]]:
+        """A ``set`` containing all key ids and Fingerprint (if any) which are indicated to have signed or encrypted this message."""
         return self.encrypters | self.signers
 
     @property
@@ -937,7 +937,7 @@ class PGPMessage(Armorable, PGPObject):
         return list(self._signatures)
 
     @property
-    def signers(self):
+    def signers(self) -> Set[Union[KeyID,Fingerprint]]:
         """A ``set`` containing all key ids (if any) which have signed this message."""
         return set(m.signer for m in self._signatures)