+++ docs/source/changelog.rst
@@ -12,10 +12,11 @@ v0.7.0
 API changes
 -----------
 
-The following property of PGPSignature now returns None if the
-corresponding subpacket is not present (it used to return an empty
+The following properties of PGPSignature now return None if the
+corresponding subpacket is not present (they used to return an empty
 string in that case):
 
+* signer
 * signer_fingerprint
 
 v0.6.0